[DROOLS-6610] remove duplications in FactHandle factories
