+++ drools-core/src/main/java/org/drools/core/common/AbstractFactHandleFactory.java
@@ -23,13 +23,12 @@ import java.util.concurrent.atomic.AtomicLong;
 
 import org.drools.core.WorkingMemoryEntryPoint;
 import org.drools.core.reteoo.ObjectTypeConf;
+import org.drools.core.rule.TypeDeclaration;
 import org.drools.core.spi.FactHandleFactory;
 
 import static java.util.stream.Collectors.toCollection;
 
-public abstract class AbstractFactHandleFactory
-    implements
-    FactHandleFactory  {
+public abstract class AbstractFactHandleFactory implements FactHandleFactory  {
 
     /** The fact id. */
     private IdsGenerator idGen;
@@ -79,15 +78,47 @@ public abstract class AbstractFactHandleFactory
                               wmEntryPoint );
     }
 
-    /* (non-Javadoc)
-     * @see org.drools.core.spi.FactHandleFactory#newFactHandle(long)
-     */
-    public abstract InternalFactHandle newFactHandle(long id,
-                                                     Object object,
-                                                     long recency,
-                                                     ObjectTypeConf conf,
-                                                     ReteEvaluator reteEvaluator,
-                                                     WorkingMemoryEntryPoint wmEntryPoint );
+    public final InternalFactHandle newFactHandle( final long id,
+                                                   final Object object,
+                                                   final long recency,
+                                                   final ObjectTypeConf conf,
+                                                   final ReteEvaluator reteEvaluator,
+                                                   final WorkingMemoryEntryPoint wmEntryPoint ) {
+        WorkingMemoryEntryPoint entryPoint = getWmEntryPoint(reteEvaluator, wmEntryPoint);
+        boolean isTrait = conf != null && conf.isTrait();
+
+        if ( conf != null && conf.isEvent() ) {
+            TypeDeclaration type = conf.getTypeDeclaration();
+            long timestamp;
+            if ( type != null && type.getTimestampExtractor() != null ) {
+                timestamp = type.getTimestampExtractor().getLongValue( reteEvaluator, object );
+            } else {
+                timestamp = reteEvaluator.getTimerService().getCurrentTime();
+            }
+            long duration = 0;
+            if ( type != null && type.getDurationExtractor() != null ) {
+                duration = type.getDurationExtractor().getLongValue( reteEvaluator, object );
+            }
+            return createEventFactHandle(id, object, recency, entryPoint, isTrait, timestamp, duration);
+        } else {
+            return createDefaultFactHandle(id, object, recency, entryPoint, isTrait);
+        }
+    }
+
+    protected DefaultFactHandle createDefaultFactHandle(long id, Object object, long recency, WorkingMemoryEntryPoint entryPoint, boolean isTrait) {
+        return new DefaultFactHandle(id, object, recency, entryPoint, isTrait);
+    }
+
+    protected EventFactHandle createEventFactHandle(long id, Object object, long recency, WorkingMemoryEntryPoint entryPoint, boolean isTrait, long timestamp, long duration) {
+        return new EventFactHandle(id, object, recency, timestamp, duration, entryPoint, isTrait);
+    }
+
+    protected WorkingMemoryEntryPoint getWmEntryPoint(ReteEvaluator reteEvaluator, WorkingMemoryEntryPoint wmEntryPoint) {
+        if (wmEntryPoint != null) {
+            return wmEntryPoint;
+        }
+        return reteEvaluator != null ? reteEvaluator.getDefaultEntryPoint() : null;
+    }
 
     public final void increaseFactHandleRecency(final InternalFactHandle factHandle) {
         factHandle.setRecency( getNextRecency() );
@@ -97,9 +128,6 @@ public abstract class AbstractFactHandleFactory
         factHandle.invalidate();
     }
 
-    /* (non-Javadoc)
-     * @see org.drools.core.spi.FactHandleFactory#newInstance()
-     */
     public abstract FactHandleFactory newInstance();
 
     public long getNextId() {
@@ -117,7 +145,7 @@ public abstract class AbstractFactHandleFactory
     public long getRecency() {
         return this.counter.get();
     }
-    
+
     public void clear(long id, long counter) {
         this.idGen = new IdsGenerator( id );
         this.counter = new AtomicLong( counter );
