+++ drools-core/src/main/java/org/drools/core/reteoo/ReteooFactHandleFactory.java
@@ -21,10 +21,6 @@ import java.io.Serializable;
 import org.drools.core.WorkingMemoryEntryPoint;
 import org.drools.core.common.AbstractFactHandleFactory;
 import org.drools.core.common.DefaultFactHandle;
-import org.drools.core.common.EventFactHandle;
-import org.drools.core.common.InternalFactHandle;
-import org.drools.core.common.ReteEvaluator;
-import org.drools.core.rule.TypeDeclaration;
 import org.drools.core.spi.FactHandleFactory;
 
 public class ReteooFactHandleFactory extends AbstractFactHandleFactory implements Serializable {
@@ -35,72 +31,21 @@ public class ReteooFactHandleFactory extends AbstractFactHandleFactory implement
         super();
     }
 
-    public ReteooFactHandleFactory(long id,
-                                   long counter) {
-        super( id,
-               counter );
-    }
-
-    /* (non-Javadoc)
-     * @see org.kie.reteoo.FactHandleFactory#newFactHandle(long)
-     */
-    public InternalFactHandle newFactHandle( final long id,
-                                             final Object object,
-                                             final long recency,
-                                             final ObjectTypeConf conf,
-                                             final ReteEvaluator reteEvaluator,
-                                             final WorkingMemoryEntryPoint wmEntryPoint ) {
-        if ( conf != null && conf.isEvent() ) {
-            TypeDeclaration type = conf.getTypeDeclaration();
-            long timestamp;
-            if ( type != null && type.getTimestampExtractor() != null ) {
-                timestamp = type.getTimestampExtractor().getLongValue( reteEvaluator, object );
-            } else {
-                timestamp = reteEvaluator.getTimerService().getCurrentTime();
-            }
-            long duration = 0;
-            if ( type != null && type.getDurationExtractor() != null ) {
-                duration = type.getDurationExtractor().getLongValue( reteEvaluator, object );
-            }
-            return new EventFactHandle( id,
-                                        object,
-                                        recency,
-                                        timestamp,
-                                        duration,
-                                        getWmEntryPoint(reteEvaluator, wmEntryPoint),
-                                        conf != null && conf.isTrait() );
-        } else {
-            return new DefaultFactHandle( id,
-                                          object,
-                                          recency,
-                                          getWmEntryPoint(reteEvaluator, wmEntryPoint),
-                                          conf != null && conf.isTrait() );
-        }
-    }
-
-    protected WorkingMemoryEntryPoint getWmEntryPoint(ReteEvaluator reteEvaluator, WorkingMemoryEntryPoint wmEntryPoint) {
-        if (wmEntryPoint != null) {
-            return wmEntryPoint;
-        }
-        return reteEvaluator != null ? reteEvaluator.getDefaultEntryPoint() : null;
+    public ReteooFactHandleFactory(long id, long counter) {
+        super( id, counter );
     }
 
     @Override
-    public DefaultFactHandle createDefaultFactHandle(long id, Object initialFact, long recency, WorkingMemoryEntryPoint wmEntryPoint) {
-        return new DefaultFactHandle(0, initialFact, 0, wmEntryPoint );
+    public DefaultFactHandle newInitialFactHandle(WorkingMemoryEntryPoint wmEntryPoint) {
+        return new DefaultFactHandle(0, InitialFactImpl.getInstance(), 0, wmEntryPoint);
     }
 
-    /* (non-Javadoc)
-     * @see org.kie.reteoo.FactHandleFactory#newInstance()
-     */
     public FactHandleFactory newInstance() {
         return new ReteooFactHandleFactory();
     }
 
-    public FactHandleFactory newInstance(long id,
-                                         long counter) {
-        return new ReteooFactHandleFactory( id,
-                                            counter );
+    public FactHandleFactory newInstance(long id, long counter) {
+        return new ReteooFactHandleFactory( id, counter );
     }
 
     public Class getFactHandleType() {
