+++ drools-core/src/main/java/org/drools/core/spi/FactHandleFactory.java
@@ -19,7 +19,6 @@ package org.drools.core.spi;
 import java.util.Collection;
 
 import org.drools.core.WorkingMemoryEntryPoint;
-import org.drools.core.common.DefaultFactHandle;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.common.ReteEvaluator;
 import org.drools.core.reteoo.ObjectTypeConf;
@@ -45,10 +44,7 @@ public interface FactHandleFactory {
                                      ReteEvaluator reteEvaluator,
                                      WorkingMemoryEntryPoint wmEntryPoint );
 
-    DefaultFactHandle createDefaultFactHandle(final long id,
-                                              final Object object,
-                                              final long recency,
-                                              final WorkingMemoryEntryPoint wmEntryPoint);
+    InternalFactHandle newInitialFactHandle(WorkingMemoryEntryPoint wmEntryPoint);
     
     /**
      * Increases the recency of the FactHandle
