+++ drools-traits/src/main/java/org/drools/traits/core/reteoo/TraitFactHandleFactory.java
@@ -18,20 +18,16 @@ package org.drools.traits.core.reteoo;
 
 import org.drools.core.WorkingMemoryEntryPoint;
 import org.drools.core.common.DefaultFactHandle;
-import org.drools.core.common.EventFactHandle;
-import org.drools.core.common.InternalFactHandle;
-import org.drools.core.common.ReteEvaluator;
-import org.drools.core.reteoo.ObjectTypeConf;
+import org.drools.core.reteoo.InitialFactImpl;
 import org.drools.core.reteoo.ReteooFactHandleFactory;
-import org.drools.core.rule.TypeDeclaration;
 import org.drools.core.spi.FactHandleFactory;
 import org.drools.traits.core.common.TraitDefaultFactHandle;
 
 public class TraitFactHandleFactory extends ReteooFactHandleFactory {
 
     @Override
-    public DefaultFactHandle createDefaultFactHandle(long id, Object initialFact, long recency, WorkingMemoryEntryPoint wmEntryPoint) {
-        return new TraitDefaultFactHandle(id, initialFact, recency, wmEntryPoint);
+    public DefaultFactHandle newInitialFactHandle(WorkingMemoryEntryPoint wmEntryPoint) {
+        return new TraitDefaultFactHandle(0, InitialFactImpl.getInstance(), 0, wmEntryPoint);
     }
 
     @Override
@@ -39,40 +35,8 @@ public class TraitFactHandleFactory extends ReteooFactHandleFactory {
         return new TraitFactHandleFactory();
     }
 
-    /* (non-Javadoc)
-     * @see org.kie.reteoo.FactHandleFactory#newFactHandle(long)
-     */
-    public InternalFactHandle newFactHandle(final long id,
-                                            final Object object,
-                                            final long recency,
-                                            final ObjectTypeConf conf,
-                                            final ReteEvaluator reteEvaluator,
-                                            final WorkingMemoryEntryPoint wmEntryPoint ) {
-        if ( conf != null && conf.isEvent() ) {
-            TypeDeclaration type = conf.getTypeDeclaration();
-            long timestamp;
-            if ( type != null && type.getTimestampExtractor() != null ) {
-                timestamp = type.getTimestampExtractor().getLongValue( reteEvaluator, object );
-            } else {
-                timestamp = reteEvaluator.getTimerService().getCurrentTime();
-            }
-            long duration = 0;
-            if ( type != null && type.getDurationExtractor() != null ) {
-                duration = type.getDurationExtractor().getLongValue( reteEvaluator, object );
-            }
-            return new EventFactHandle(id,
-                                       object,
-                                       recency,
-                                       timestamp,
-                                       duration,
-                                       wmEntryPoint != null ? wmEntryPoint : reteEvaluator.getDefaultEntryPoint(),
-                                       conf != null && conf.isTrait() );
-        } else {
-            return new TraitDefaultFactHandle(id,
-                                              object,
-                                              recency,
-                                              wmEntryPoint != null ? wmEntryPoint : reteEvaluator.getDefaultEntryPoint(),
-                                              conf != null && conf.isTrait() );
-        }
+    @Override
+    protected DefaultFactHandle createDefaultFactHandle(long id, Object object, long recency, WorkingMemoryEntryPoint entryPoint, boolean isTrait) {
+        return new TraitDefaultFactHandle(id, object, recency, entryPoint, isTrait);
     }
 }
